pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-chimera"
version = "0.1.0"
description = "Project Chimera - Enhanced Developer's Toolkit with structured MCP separation"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Chimera Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.1,<0.115.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "psutil>=5.9.6",
    "pygments>=2.16.1",
    "black>=23.10.1",
    "pylint>=3.0.2",
    "gitpython>=3.1.40",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2,<3.0.0",
    "typing-extensions>=4.8.0",
    "starlette>=0.27.0,<0.46.0",
    "click>=8.1.7",
    "h11>=0.14.0",
    "idna>=3.4",
    "sniffio>=1.3.0",
    "websockets>=11.0.3",
    "mcp>=1.5.0,<2.0.0",
    "httpx-sse>=0.4.0",
    "httpx>=0.24.1",
    "sse-starlette>=2.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["chimera_core", "chimera_data", "chimera_stdio_mcp", "chimera_ui", "schemas"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 